Chnage double issue

+++ lib/inprogress/ProfileReview.dart
@@ -214,7 +214,7 @@ class _ProfileReviewState extends State<ProfileReview> {
                       USER_DOC_SiGN_URL, data.userData!.docSignedUrl!
                   );
                 }
-                prefsUtil.saveDouble(USER_PAY_OUT, data.userData!.payout!);
+                prefsUtil.saveDouble(USER_PAY_OUT, data.userData!.payout!.toDouble());
                 if( data.userData!.docSignedUrl!=null) {
                   prefsUtil.saveString(
                       USER_DOC_SiGN_URL, data.userData!.docSignedUrl!);
