+++ lib/view/splash/splash_screen.dart
@@ -172,7 +172,7 @@ class _SplashScreenState extends State<SplashScreen> {
                       USER_DOC_SiGN_URL, data.userData!.docSignedUrl!
                   );
                 }
-                prefsUtil.saveDouble(USER_PAY_OUT, data.userData!.payout!);
+                prefsUtil.saveDouble(USER_PAY_OUT, data.userData!.payout!.toDouble());
                 if (data.userData!.docSignedUrl != null) {
                   prefsUtil.saveString(
                       USER_DOC_SiGN_URL, data.userData!.docSignedUrl!);
